update dtos to match rac v2 structure
also reorganize all dtos into /dto folder and suffix filenames
with "-dto"
