+++ src/dto/allocation-dto.ts
@@ -1,0 +1,6 @@
+import { IVariation } from './variation-dto';
+
+export interface IAllocation {
+  percentExposure: number;
+  variations: IVariation[];
+}
