+++ src/dto/variation-dto.ts
@@ -5,5 +5,7 @@ export interface IShardRange {
 
 export interface IVariation {
   name: string;
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  value: any;
   shardRange: IShardRange;
 }
